[fix] 팝업 - 미화원 로티 연결
